Add different Levels, Add Result Screen, Add Multiplayer, Refine Pausing
Added Multiplayer
-------------------

Player now has an index tied to the game controller that controls them. The index is passed into their input component and stored as an attribute there

level.js Listens for gamepads connected, and creates new players of the equal index to gamepad

If a gamepad is disconnected the associated player is removed.

input.js no longer listens for gamepads connected as level.js fufills this

Get gamepad now tries to get gamepad at the input.js specified index, but will not crash if it fails.

Added Resources
-------------------

An attempt was made to add an animator class, but this failed. The code is kept in tact for viewing.

Image resources were refined to hold arrays of images (frames of animation)

Each player has a separate set of resources. In the player constructer, the default resources correspond to the index found in resources.js

Resources.js defines a path for every object found in it, meaning different players and enemies all follow the same basic layout.

Add levels
-------------
There are now multiple levels in level.js

You can select a level by passing it into the constructer

Level 1 and 2 are predesigned levels. Level 3 is procedurally generated, meaning. Once you finish 1 and 2, the game will continue generating new levels.

Level 3's collectibles and enemies will always spawn above platforms. The player will also always spawn on a platform.

loadLevel() lets you load any level given an int, allowing for easy progression.

Refine Pausing
----------------
When the game is paused, every player's input is disabled. This means that the game appears truly motionless as opposed to player sprites flipping even while paused.

Gameplay
------------
Any player who dies will now stay dead. Their renderer and physics will be disabled, meaning they effectively can't move.

Player
-----------
Increased the player's speed from 100 to 200

The player's spawn point is now stored as it will be modified for ResetPlayerState

The player triggers jumping audio

Location.reload is no longer utilised for resetting the game.

If the player reaches the level 1 or 2's goal they will trigger the next level.

Player score and player state is now reset independently

Reset player state also resets the player's speed and jumpforce as collectibles may modify these.

Collectibles
-------------
All collectibles contribute to score now

Camera
---------
In order to focus on the center of the screen, an empty was created. The empty did not have a renderer, but the camera needed a renderer.

This requirement was removed, and the camera may now focus on any object.

A try catch was added as a fail safe, in case it trys to follow an object which no longer exists.

Results Screen
----------------
The camera is focused on the center of the screen. Each player is placed on their own platform, with their lives, score, and whether they won displayed over their head.

This screen is displayed for 5 seconds, then exits

playerUI
-----------
playerUI now has two UI components. A winnerUI component lies directly above the scoreUI. It is disabled by default, and is only enabled when a player wins.

A new function resultsText centers the text for the results screen. gameText returns it to its original position.

As there are now multiple players, playerUI now takes a player in its constructor to keep track of different scores and lives.

Game Management
--------------
These aspects will be moved to their own game Manager object. For now, they are in levels.

When all collectibles have been collected, in Level 1 and 2 new collectibles will be generated. Otherwise, in level 3, a new level will be generated.

If the player who the camera is focused on dies, a new player will be chosen.

If all players are dead, the game will skip to results screen

The contents of update in game.js have been moved to level.js, this will be undone.