Add Games Manager, refine multiplayer
levels.js
--------------
Anything not inherently related to the level was moved to game.js or gameManager.js (or anything that requires an update() method)

If there are no winners, set the camera to focus on a random player

Information on players and playerUIs is now taken from gameManager.

game.js
---------
All game states are now stored here (isPaused, viewingResults, inLevel)

Controller connection and disconnection is listened to from this script.

gameManager
----------------

Added gameManager, responsible for switching gameplay states and handling keeping track of and adding or removing players.

The game manager is a singleton, similar to the audioManager. Similar to camera, it does not inherit from any class and its update is called from game.js

Any code in level.js related to game management was adjusted and moved here.

Adjusted refocusCamera so the conditions are not built into the function

Adjusted onPlayersDead so it checks whether there are any players in the scene first.

Added tallyScores and setWinners which replace otherwise irrelevent scripting in level.js

Added AddPlayer()
The index is the player's number and controller which controls the player.

isCameraTarget determines whether the camera will follow the player. This is only used at the start of the game in level.js, or if there are no players currently in the game

If the index matches an existing player, or there are too many player (3) in the game, they will be skipped

If a player tries to join in the middle of viewing results or paused, they will be added to a queue. Once in a setting where players can join they will be added.

If none of these conditions apply, the player will simply be added.

Added removePlayer() which is called whenever a controller is disconnected.

It removes the player and their associated UI.

If the player was the target of the camera previously, the camera will be refocused